# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "QRouter API",
    "version": "1.0.0",
    "description": "One authenticated API for hardware-aware compilation, routing, pricing, and execution across quantum providers."
  },
  "servers": [
    {
      "url": "https://api.qrouter.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "qci_live_*"
      }
    },
    "parameters": {
      "JobId": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "ProjectId": {
        "in": "path",
        "name": "id",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "schemas": {
      "Constraints": {
        "type": "object",
        "properties": {
          "maxCost": {
            "type": "number",
            "exclusiveMinimum": 0
          },
          "maxQueueSeconds": {
            "type": "integer",
            "minimum": 0
          },
          "minFidelity": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "kind": {
            "enum": [
              "qpu",
              "simulator"
            ]
          },
          "providers": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "excludeProviders": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateJob": {
        "type": "object",
        "required": [
          "circuit"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "circuit": {
            "type": "string",
            "maxLength": 256000
          },
          "format": {
            "enum": [
              "openqasm2",
              "openqasm3"
            ],
            "default": "openqasm2"
          },
          "shots": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000,
            "default": 1024
          },
          "target": {
            "type": "string",
            "default": "auto"
          },
          "routing_mode": {
            "enum": [
              "balanced",
              "cost",
              "speed",
              "quality"
            ],
            "default": "balanced"
          },
          "optimization_level": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3,
            "default": 2
          },
          "failover": {
            "type": "boolean",
            "default": true,
            "description": "Try another compatible backend after a provider failure without exceeding the accepted quote."
          },
          "max_attempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 3
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 604800,
            "default": 7200
          },
          "constraints": {
            "$ref": "#/components/schemas/Constraints"
          }
        }
      },
      "ProjectSettings": {
        "type": "object",
        "required": [
          "shots",
          "target",
          "routingMode",
          "optimizationLevel"
        ],
        "properties": {
          "shots": {
            "type": "integer",
            "minimum": 1,
            "maximum": 1000000,
            "default": 1024
          },
          "target": {
            "type": "string",
            "default": "auto"
          },
          "routingMode": {
            "enum": [
              "balanced",
              "cost",
              "speed",
              "quality"
            ],
            "default": "balanced"
          },
          "optimizationLevel": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3,
            "default": 2
          },
          "failover": {
            "type": "boolean",
            "default": true
          },
          "maxAttempts": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "default": 3
          },
          "timeoutSeconds": {
            "type": "integer",
            "minimum": 60,
            "maximum": 604800,
            "default": 7200
          }
        }
      },
      "CreateProject": {
        "type": "object",
        "required": [
          "repository"
        ],
        "properties": {
          "repository": {
            "type": "string",
            "examples": [
              "acme/quantum-workloads"
            ]
          },
          "production_branch": {
            "type": "string",
            "default": "repository default branch"
          },
          "circuit_path": {
            "type": "string",
            "examples": [
              "circuits/bell.qasm"
            ]
          },
          "settings": {
            "$ref": "#/components/schemas/ProjectSettings"
          }
        }
      },
...
      "Job": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "enum": [
              "created",
              "analyzing",
              "quoted",
              "awaiting_payment",
              "funds_reserved",
              "queued",
              "dispatching",
              "submitted",
              "processing",
              "completed",
              "failed",
              "cancellation_requested",
              "cancelled"
            ]
          },
          "selected_backend_id": {
            "type": "string"
          },
          "analysis": {
            "type": "object"
          },
          "route_decision": {
            "type": "object"
          },
          "result": {
            "type": [
              "object",
              "null"
            ]
          },
          "quote": {
            "type": "object",
            "nullable": true,
            "description": "Quote issued for this job (present on creation and detail responses).",
            "properties": {
              "providerCost": {
                "type": "number"
              },
              "transpilerFee": {
                "type": "number"
              },
              "platformFee": {
                "type": "number"
              },
              "total": {
                "type": "number"
              },
              "currency": {
                "type": "string"
              },
              "expiresAt": {
                "type": "string",
                "format": "date-time"
              },
              "rateSnapshot": {
                "type": "object"
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "summary": "List connected repository projects",
        "responses": {
          "200": {
            "description": "Project list"
          }
        }
      },
      "post": {
        "summary": "Import or update a repository project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProject"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Project imported"
          },
          "400": {
            "description": "Invalid project configuration"
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "patch": {
        "summary": "Update production source and routing defaults",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "200": {
            "description": "Project updated"
          },
          "404": {
            "description": "Project not found"
          }
        }
      },
      "delete": {
        "summary": "Remove a repository project",
        "parameters": [
          {
            "$ref": "#/components/parameters/ProjectId"
          }
        ],
        "responses": {
          "204": {
            "description": "Project removed"
          }
        }
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "summary": "List jobs",
        "responses": {
          "200": {
            "description": "Job list"
          }
        }
      },
      "post": {
        "summary": "Compile, quote, route, and durably queue a circuit",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJob"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Demo job created"
          },
          "202": {
            "description": "Job accepted into the durable execution queue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "402": {
            "description": "Insufficient credits"
          },
          "422": {
            "description": "Invalid circuit or no compatible backend"
          },
          "503": {
            "description": "Compiler unavailable"
          }
        }
      }
    },
...
    "/api/v1/jobs/{id}/cancel": {
      "post": {
        "summary": "Cancel queued work and release reserved credits",
        "parameters": [
          {
            "$ref": "#/components/parameters/JobId"
          }
        ],
        "responses": {
          "200": {
            "description": "Cancelled"
          },
          "409": {
            "description": "Already processing or terminal"
          }
        }
      }
    },
...
  }
}