# src/app/api/v1/jobs/route.ts
import { createHash, randomUUID } from "crypto";
import { NextResponse } from "next/server";
import { analyzeCircuit } from "@/lib/qrouter/analyze";
import { storeArtifact } from "@/lib/qrouter/artifacts";
import { resolvePrincipal } from "@/lib/qrouter/auth";
import { demoJobs, type StoredJob } from "@/lib/qrouter/demo-store";
import { submitToProvider } from "@/lib/qrouter/execution";
import { apiError } from "@/lib/qrouter/http";
import { prepareExecution } from "@/lib/qrouter/pipeline";
import { loadRoutingContext } from "@/lib/qrouter/routingContext";
import { publicTranspilation } from "@/lib/qrouter/transpiler";
import { createJobSchema } from "@/lib/qrouter/validation";
import { createAdminClient } from "@/lib/supabase/admin";

export const dynamic = "force-dynamic";

export async function GET(request: Request) {
  try {
    const principal = await resolvePrincipal(request);
    if (principal.demo) {
      const data = [...demoJobs.values()].filter((job) => job.organization_id === principal.organizationId).sort((a, b) => b.created_at.localeCompare(a.created_at));
      return NextResponse.json({ object: "list", data });
    }
    const admin = createAdminClient();
    const { data, error } = await admin.from("jobs").select("id,project_id,request_id,name,input_format,shots,target,routing_mode,status,selected_backend_id,failover_enabled,max_attempts,execution_deadline_at,analysis,route_decision,result,error,created_at,updated_at,completed_at").eq("organization_id", principal.organizationId).order("created_at", { ascending: false }).limit(100);
    if (error) throw error;
    return NextResponse.json({ object: "list", data });
  } catch (error) {
    return apiError(error);
  }
}

export async function POST(request: Request) {
  try {
    const principal = await resolvePrincipal(request);
    const parsed = createJobSchema.safeParse(await request.json());
    if (!parsed.success) {
      return NextResponse.json({ error: { type: "invalid_request", message: "The request body is invalid.", details: parsed.error.flatten() } }, { status: 400 });
    }
    const input = parsed.data;
    const originalAnalysis = analyzeCircuit(input.circuit, input.format);
    const { snapshot, backends } = await loadRoutingContext(principal.demo);
    const prepared = await prepareExecution({
      backends,
      analysis: originalAnalysis,
      shots: input.shots,
      target: input.target,
      mode: input.routing_mode,
      constraints: input.constraints,
      qciSnapshotId: snapshot.id,
      qciTimestamp: snapshot.ts,
      optimizationLevel: input.optimization_level,
    });
    const { decision, quote, executionAnalysis } = prepared;
    const analysis = { ...originalAnalysis, transpilation: publicTranspilation(prepared.transpilation) };
    const idempotencyKey = request.headers.get("idempotency-key")?.trim() || null;
    const requestId = request.headers.get("x-request-id")?.trim() || randomUUID();
    const now = new Date().toISOString();
    const jobId = randomUUID();

    if (principal.demo) {
      if (idempotencyKey) {
        const existing = [...demoJobs.values()].find((job) => job.organization_id === principal.organizationId && (job as StoredJob & { idempotency_key?: string }).idempotency_key === idempotencyKey);
        if (existing) return NextResponse.json(existing);
      }
      const base: StoredJob & { idempotency_key?: string } = {
        id: jobId, organization_id: principal.organizationId, name: input.name ?? null,
        input_format: input.format, source: input.circuit, shots: input.shots, target: input.target,
        routing_mode: input.routing_mode, status: "submitted", selected_backend_id: decision.selected.id,
        analysis, route_decision: decision, quote, result: null, error: null,
        created_at: now, updated_at: now, completed_at: null, idempotency_key: idempotencyKey ?? undefined,
      };
      demoJobs.set(jobId, base);
      try {
        const submission = await submitToProvider(decision.selected.id, executionAnalysis, input.shots, jobId);
        base.status = submission.status === "completed" ? "completed" : "submitted";
        base.result = submission.result ?? null;
        base.updated_at = new Date().toISOString();
        base.completed_at = submission.status === "completed" ? base.updated_at : null;
      } catch (error) {
        base.status = "failed";
        base.error = { message: error instanceof Error ? error.message : "Execution failed." };
      }
      return NextResponse.json(base, { status: 201 });
    }

    const admin = createAdminClient();
    if (idempotencyKey) {
      const { data: existing } = await admin.from("jobs").select("*").eq("organization_id", principal.organizationId).eq("idempotency_key", idempotencyKey).maybeSingle();
      // Terminal failures release their idempotency key so the same deployment
      // (e.g. a repository commit) can be retried after the cause is fixed.
      if (existing && !["failed", "cancelled"].includes(existing.status as string)) return NextResponse.json(existing);
      if (existing) {
        const { error: unpinError } = await admin.from("jobs").update({ idempotency_key: null, updated_at: new Date().toISOString() }).eq("id", existing.id);
        if (unpinError) throw unpinError;
      }
    }
    const sourceHash = createHash("sha256").update(input.circuit).digest("hex");
    const { error: jobError } = await admin.from("jobs").insert({
      id: jobId, organization_id: principal.organizationId, user_id: principal.userId,
      api_key_id: principal.apiKeyId, idempotency_key: idempotencyKey, name: input.name,
      request_id: requestId,
      input_format: input.format, source: input.circuit, source_hash: sourceHash, shots: input.shots,
      target: input.target, routing_mode: input.routing_mode, constraints: input.constraints,
      analysis, route_decision: decision, selected_backend_id: decision.selected.id, status: "quoted",
      failover_enabled: input.failover, max_attempts: input.max_attempts, execution_timeout_seconds: input.timeout_seconds, next_attempt_at: now,
    });
    if (jobError) throw jobError;
    await Promise.all([
      storeArtifact({ jobId, organizationId: principal.organizationId, kind: "source", content: input.circuit }),
      storeArtifact({
        jobId,
        organizationId: principal.organizationId,
        kind: "transpiled",
        content: prepared.transpilation.artifactQasm ?? prepared.transpilation.qasm,
      }),
    ]);
    // Quote storage, credit reservation, and the queued/awaiting_payment
    // transition happen in one transaction so a crash cannot strand the job in
    // a half-transitioned state or leak a reservation.
    const { data: outcome, error: queueError } = await admin.rpc("queue_job_with_quote", {
      p_job_id: jobId,
      p_provider_cost: quote.providerCost,
      p_transpiler_fee: quote.transpilerFee,
      p_platform_fee: quote.platformFee,
      p_total: quote.total,
      p_rate_snapshot: quote.rateSnapshot,
      p_expires_at: quote.expiresAt,
    });
    if (queueError) throw queueError;
    if (outcome === "awaiting_payment") {
      return NextResponse.json({ error: { type: "insufficient_credits", message: "Add credits before running this task.", quote, job_id: jobId } }, { status: 402 });
    }
    const { data: created, error } = await admin.from("jobs").select("*").eq("id", jobId).single();
    if (error) throw error;
    return NextResponse.json({ ...created, quote }, { status: 202 });
  } catch (error) {
    return apiError(error);
  }
}